introduce pickly notifications api
